requirements.txt
autopep8==1.5.3
backcall==0.2.0
certifi==2020.4.5.2
chardet==3.0.4
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
Django==2.1.15
django-allauth==0.42.0
django-bootstrap-pagination==1.7.1
django-cors-headers==3.3.0
django-decouple==2.1
django-extensions==2.2.9
django-oauth-toolkit==1.3.2
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
oauthlib==3.1.0
parso==0.7.0
pickleshare==0.7.5
Pillow==7.1.2
prompt-toolkit==3.0.5
pycodestyle==2.6.0
Pygments==2.6.1
PyJWT==1.7.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
toml==0.10.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.4
